+++ examples/daemons/install_hls.sh
@@ -5,7 +5,7 @@ echo "
 Description=HiddenLakeService
 
 [Service]
-ExecStart=/root/hls_linux -path=/root
+ExecStart=/root/hls_amd64_linux -path=/root
 Restart=always
 RestartSec=10
 
@@ -14,9 +14,9 @@ WantedBy=multi-user.target
 " > /etc/systemd/system/hidden_lake.service
 
 chdir /root
-wget https://github.com/number571/go-peer/releases/download/v1.5.11/hls_linux
+wget https://github.com/number571/go-peer/releases/download/v1.5.11/hls_amd64_linux
 
-chmod +x /root/hls_linux
+chmod +x /root/hls_amd64_linux
 systemctl daemon-reload
 systemctl enable hidden_lake.service
 systemctl restart hidden_lake.service