+++ pkg/network/conn/settings.go
@@ -11,6 +11,8 @@ type sSettings struct {
 	FNetworkKey    string
 	FMessageSize   uint64
 	FLimitVoidSize uint64
+	FReadDeadline  time.Duration
+	FWriteDeadline time.Duration
 	FFetchTimeWait time.Duration
 }
 
@@ -19,6 +21,8 @@ func NewSettings(pSett *SSettings) ISettings {
 		FNetworkKey:    pSett.FNetworkKey,
 		FMessageSize:   pSett.FMessageSize,
 		FLimitVoidSize: pSett.FLimitVoidSize,
+		FReadDeadline:  pSett.FReadDeadline,
+		FWriteDeadline: pSett.FWriteDeadline,
 		FFetchTimeWait: pSett.FFetchTimeWait,
 	}).mustNotNull()
 }
@@ -29,6 +33,12 @@ func (p *sSettings) mustNotNull() ISettings {
 	}
 	if p.FLimitVoidSize == 0 {
 		panic(`p.FMaxVoidSize == 0`)
+	}
+	if p.FReadDeadline == 0 {
+		panic(`p.FReadDeadline == 0`)
+	}
+	if p.FWriteDeadline == 0 {
+		panic(`p.FWriteDeadline == 0`)
 	}
 	if p.FFetchTimeWait == 0 {
 		panic(`p.FFetchTimeWait == 0`)
@@ -46,6 +56,14 @@ func (p *sSettings) GetMessageSize() uint64 {
 
 func (p *sSettings) GetLimitVoidSize() uint64 {
 	return p.FLimitVoidSize
+}
+
+func (p *sSettings) GetReadDeadline() time.Duration {
+	return p.FReadDeadline
+}
+
+func (p *sSettings) GetWriteDeadline() time.Duration {
+	return p.FWriteDeadline
 }
 
 func (p *sSettings) GetFetchTimeWait() time.Duration {