+++ pkg/network/settings.go
@@ -12,20 +12,20 @@ var (
 
 type SSettings sSettings
 type sSettings struct {
-	FAddress       string
-	FCapacity      uint64
-	FMaxConnects   uint64
-	FActionTimeout time.Duration
-	FConnSettings  conn.ISettings
+	FAddress      string
+	FCapacity     uint64
+	FMaxConnects  uint64
+	FWriteTimeout time.Duration
+	FConnSettings conn.ISettings
 }
 
 func NewSettings(pSett *SSettings) ISettings {
 	return (&sSettings{
-		FAddress:       pSett.FAddress,
-		FCapacity:      pSett.FCapacity,
-		FMaxConnects:   pSett.FMaxConnects,
-		FActionTimeout: pSett.FActionTimeout,
-		FConnSettings:  pSett.FConnSettings,
+		FAddress:      pSett.FAddress,
+		FCapacity:     pSett.FCapacity,
+		FMaxConnects:  pSett.FMaxConnects,
+		FWriteTimeout: pSett.FWriteTimeout,
+		FConnSettings: pSett.FConnSettings,
 	}).mustNotNull()
 }
 
@@ -36,8 +36,8 @@ func (p *sSettings) mustNotNull() ISettings {
 	if p.FMaxConnects == 0 {
 		panic(`p.FMaxConnects == 0`)
 	}
-	if p.FActionTimeout == 0 {
-		panic(`p.FActionTimeout == 0`)
+	if p.FWriteTimeout == 0 {
+		panic(`p.FWriteTimeout == 0`)
 	}
 	if p.FConnSettings == nil {
 		panic(`p.FConnSettings == nil`)
@@ -61,6 +61,6 @@ func (p *sSettings) GetConnSettings() conn.ISettings {
 	return p.FConnSettings
 }
 
-func (p *sSettings) GetActionTimeout() time.Duration {
-	return p.FActionTimeout
+func (p *sSettings) GetWriteTimeout() time.Duration {
+	return p.FWriteTimeout
 }