+++ CHANGELOG.md
@@ -22,6 +22,7 @@ In progress...
 
 - Makefule `hidden_lake`: _std_cfg -> _std_cfgs
 - Package `network`: fix deadlock BroadcastPayload on WritePayload method
+- Package `network/conn`: append read/write deadline methods
 
 ## v1.5.11